+++ js/mock-data.js
@@ -1,4 +1,15 @@
-import {commentTextTemplate, authorNames} from './data.js';
+// Template names for comments
+export const authorNames = ['Александр', 'Матвей', 'Артем', 'Михаил', 'Иван', 'Тимофей', 'Кирилл', 'Дмитрий', 'Даниил', 'Мария',
+  'Варвара', 'Виктория', 'Алиса', 'Ксения', 'Василиса', 'Милана', 'Ева', 'Александра'];
+
+// Template text for comments
+export const commentTextTemplate = 'Всё отлично!\n' +
+  'В целом всё неплохо. Но не всё.\n' +
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.\n' +
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.\n' +
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.\n' +
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!';
+
 
 // Pick random number function
 export const getRandomNumber = (from, to) => {
